refactor: Date Util을 Remind Mananger로 합병 및 broadcast receiver async 작업 추가
1. goAsync()를 이용하여 pendingResult 타입을 반환받고 작업 종료후 result 종료
2. Remind Manager로 관리함으로써 Date Util을 제거할 수 있었음
3. Constant file에서 다양한 상수를 관리하였는데, constant에서 관리하는 것보다 관련된 class의 companion object 내부에 선언하는 것이 의도를 더 잘 드러낸다고 판단
4.
