+++ app/src/main/java/com/example/remindapp/MainActivity.kt
@@ -12,7 +12,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.fragment.NavHostFragment
 import com.example.remindapp.databinding.ActivityMainBinding
-import com.example.remindapp.util.REMIND_IDX
+import com.example.remindapp.util.SELECTED_REMIND_IDX
 
 class MainActivity : AppCompatActivity() {
 
@@ -43,12 +43,12 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         checkPermission()
-        val idx = intent?.extras?.get(REMIND_IDX)
+        val idx = intent?.extras?.get(SELECTED_REMIND_IDX)
 
         if (idx != null) {
             val id = idx as Int
             val bundle = Bundle()
-            bundle.putInt(REMIND_IDX, id)
+            bundle.putInt(SELECTED_REMIND_IDX, id)
             val navHostFragment =
                 supportFragmentManager.findFragmentById(R.id.nav_host_fragment_activity_main) as NavHostFragment
             val navController = navHostFragment.navController
