+++ app/src/main/java/com/example/remindapp/ui/edit/EditFragment.kt
@@ -14,7 +14,6 @@ import androidx.activity.result.ActivityResultCallback
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.net.toUri
-import androidx.core.os.bundleOf
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
@@ -24,6 +23,7 @@ import com.example.remindapp.databinding.FragmentEditBinding
 import com.example.remindapp.model.repository.RemindLocalDatasource
 import com.example.remindapp.model.repository.RemindRepository
 import com.example.remindapp.model.room.RemindDatabase
+import com.example.remindapp.ui.home.HomeFragment.Companion.SELECTION
 import com.example.remindapp.util.*
 
 class EditFragment : Fragment() {
