+++ app/src/main/java/com/example/remindapp/util/Constant.kt
@@ -1,11 +1,6 @@
 package com.example.remindapp.util
 
-const val REMIND_IDX = "remindIdx"
-const val ALARM_STATE = "alarm"
-const val SELECTION = "selection"
-const val ON = 1
-const val OFF = 0
-const val DAY_MILLIS = 86400000L
+const val SELECTED_REMIND_IDX = "remindIdx"
 const val RESULT_KEY_CHECK_CHANGE = "NotificationChange"
 const val BUNDLE_KEY_CHECK_CHANGE = "isChanged"
 const val BUNDLE_KEY_IDX = "RemindIdx"
