+++ app/src/main/java/com/example/remindapp/util/RemindManager.kt
@@ -6,6 +6,9 @@ import android.content.Context
 import android.content.Intent
 import com.example.remindapp.component.AlarmReceiver
 import com.example.remindapp.model.room.Remind
+import java.lang.Exception
+import java.text.SimpleDateFormat
+import java.util.*
 
 object RemindManager {
 
@@ -23,7 +26,7 @@ object RemindManager {
         val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
         val dayPlus = checkRemindAlarmOnNextDayOrNot(remind)
         val pending = Intent(context, AlarmReceiver::class.java).apply {
-            putExtra(REMIND_IDX, remind.id)
+            putExtra(SELECTED_REMIND_IDX, remind.id)
         }.run {
             PendingIntent.getBroadcast(
                 context,
@@ -37,7 +40,6 @@ object RemindManager {
         if (curMillis == -1L) return
 
         alarmManager.setExactAndAllowWhileIdle(AlarmManager.RTC_WAKEUP, curMillis, pending)
-        println("Result ${remind}, Result idx: ${remind.id}")
     }
 
     private fun checkRemindAlarmOnNextDayOrNot(remind: Remind): Boolean {
@@ -45,4 +47,34 @@ object RemindManager {
         return if (remind.hour < curHour) { true } else remind.hour == curHour && remind.minute <= curMin
     }
 
+    private fun getCurrentTime(): List<Int> {
+        return SimpleDateFormat("HH:mm", Locale.getDefault()).format(Date()).split(":").map { it.toInt() }
+    }
+
+    private fun getCurrentDay(): String {
+        return SimpleDateFormat("yyyy-MM-dd", Locale.getDefault()).format(Date())
+    }
+
+    private fun getTargetTime(hour: Int, minute: Int): String {
+        return "$hour:$minute:00"
+    }
+
+    private fun convertDateToMillis(hour: Int, minute: Int, flag: Boolean): Long {
+        val date = convertTime(hour, minute)
+        val pattern = "yyyy-MM-dd HH:mm:ss"
+        val parser = SimpleDateFormat(pattern, Locale.getDefault())
+        val result = try {
+            parser.parse(date).time
+        } catch (e: Exception) {
+            e.printStackTrace()
+            -1L
+        }
+        return if(flag) result + DAY_MILLIS else result
+    }
+
+    private fun convertTime(hour: Int, minute: Int): String {
+        return "${getCurrentDay()} ${getTargetTime(hour, minute)}"
+    }
+
+    private const val DAY_MILLIS = 86400000L
 }
